Refactor: create common and unpack directory and move cpps as needed. msix.cpp had to be splitted into common\msix.cpp and unpack\msixUnpack.cpp. I still need to refactor ZipObject.cpp, ideally it will end up in common, but I might have a zip writer and reader. Both will end up in common as the reader is needed for our resource zip file
