+++ .gitignore
@@ -398,7 +398,7 @@ __pycache__/
 lib/zlib
 # Files generated by CMake
 src/inc/MSIXResource.hpp
-src/msix/MSIXResource.cpp
+src/msix/common/MSIXResource.cpp
 
 .vscode/
