+++ src/msix/common/AppxManifestObject.cpp
