+++ src/msix/common/AppxPackageInfo.cpp
