+++ src/msix/common/Encoding.cpp
