+++ src/msix/unpack/AppxBundleManifest.cpp
