+++ src/msix/unpack/AppxPackageObject.cpp
